Refactored slightly and matches original output.
After some work, I’ve realised that the original algorithm produces
different output when run over multiple cores, compared to using a
single core. My optimised algorithm is producing the same output as the
original algorithm, depending on whether we are using multiple cores or
not. I think that this means the multiple cores algorithm is currently
incorrect.